+++ src/puzzle/Config.js
@@ -53,7 +53,7 @@
 				"patchwork_leftaux",
 				true
 			); /* patchwork: Alternative mouse input */
-
+			this.add("rampage_coloring", false); /* rampage: scott-pattern grid */
 			this.add("squarecell", true); /* セルは正方形にする */
 
 			/* 入力方法設定 */
@@ -402,6 +402,9 @@
 				case "patchwork_leftaux":
 					exec = pid === "patchwork";
 					break;
+				case "rampage_coloring":
+					exec = pid === "rampage";
+					break;
 				case "undefcell":
 					exec = pid === "shugaku" || pid === "lightshadow";
 					break;
@@ -498,6 +501,7 @@
 				case "disptype_interbd":
 				case "dispqnumbg":
 				case "mouseonly":
+				case "rampage_coloring":
 					puzzle.redraw();
 					break;