+++ src/variety/turnaround.js
@@ -12,8 +12,7 @@
 				"line",
 				"peke",
 				"clear",
-				"info-line",
-				"shade", "unshade"
+				"info-line"
 			]
 		},
 		autoedit_func: "qnum",
@@ -25,7 +24,53 @@
 
 	Cell: {
 		minnum: 0,
-		maxnum: 3
+		maxnum: 3,
+
+		isCmp: function() {
+			if (!this.puzzle.execConfig("autocmp")) {
+				return false;
+			}
+			return this.qnum === this.countTurn();
+		},
+
+		countTurn: function() {
+			if (!this.isNum() || this.lcnt !== 2) {
+				return -1;
+			}
+			
+			var clist = this.getAdjCells();
+			if(clist.some(function(cell) {
+				return cell.lcnt !== 2;
+			})) {
+				return -1;
+			}
+			
+			clist.add(this);
+			var turncount = 0;
+			clist.each(function(cell) {
+				turncount += cell.isLineCurve() ? 1 : 0;
+			});
+			
+			return turncount;
+		},
+
+		getAdjCells: function() {
+			var clist = new this.klass.PieceList();
+			for(var dir in this.adjborder) {
+				if (this.adjborder[dir].isLine()) {
+					clist.add(this.adjacent[dir]);
+				}
+			}
+			return clist;
+		},
+
+		posthook: {
+			qans: function(num) {
+				if(num >= 0) {
+					this.checkAutoCmp();
+				}
+			}
+		}
 	},
 	Board: {
 		hasborder: 1
@@ -35,11 +80,34 @@
 	},
 
 	Graphic: {
-		// irowake: true,
+		irowake: true,
+		qcmpcolor: "rgb(127,127,127)",
+		autocmp: "number",
+
+		getQuesNumberColor: function(cell) {
+			var qnum_color = this.getQuesNumberColor_mixed(cell);
+			if ((cell.error || cell.qinfo) === 1) {
+				return qnum_color;
+			}
+			return cell.isCmp() ? this.qcmpcolor : qnum_color;
+		},
+
+		setRange: function(x1, y1, x2, y2) {
+			var puzzle = this.puzzle,
+				bd = puzzle.board;
+			if (puzzle.execConfig("autocmp")) {
+				x1 = bd.minbx - 2;
+				y1 = bd.minby - 2;
+				x2 = bd.maxbx + 2;
+				y2 = bd.maxby + 2;
+			}
+			this.common.setRange.call(this, x1, y1, x2, y2);
+		},
+
 		paint: function() {
 			this.drawBGCells();
 			this.drawShadedCells();
-			this.drawGrid();
+			this.drawDashedGrid();
 			this.drawLines();
 			this.drawQuesNumbers();
 			this.drawPekes();
@@ -74,34 +142,26 @@
 			"checkBranchLine",
 			"checkCrossLine",
 
-			"checkTurn",
+			"checkTurnCount",
 			"checkNoLineNumber",
 
 			"checkDeadendLine+",
 			"checkOneLoop"
 		],
 
-		checkTurn: function() {
+		checkTurnCount: function() {
 			this.checkAllCell(function(cell){
-				if (!cell.isNum() || cell.lcnt !== 2) {
+				if(!cell.isValidNum()) {
 					return false;
 				}
 
-				var clist = new cell.klass.PieceList();
-				for(var d in cell.adjborder) {
-					if (cell.adjborder[d].isLine()) {
-						clist.add(cell.adjacent[d]);
-					}
+				var turncount = cell.countTurn();
+				if (turncount === -1) {
+					return false;
 				}
-
-				var turncount = cell.isLineStraight() ? 0 : 1;
-				clist.each(function(cell) {
-					turncount += cell.isLineStraight() ? 0 : 1;
-				});
-
-				if (turncount > cell.qnum || turncount < cell.qnum) {
+				if (turncount !== cell.qnum) {
 					cell.seterr(1);
-					clist.seterr(1);
+					cell.getAdjCells().seterr(1);
 					return true;
 				}